# src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: VouckBack
  description: |-
    Microserviço para geração de voucher como descontos em valor ou percentual paras clientes em seu ponto de venda.

    _Se você que saber mais sobre a TIBBI, então  click [aqui](http://tibbi.com.br)._

  contact:
    email: mario7lagoas@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:8080/voucher-back/v1
...
paths:
...
  /loja:
    get:
      tags:
        - Loja
      summary: Buscando Lista de Loja
      description: Buscando Lista de Loja na Plataforma
      operationId: buscandoListaLoja
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LojaApiResponse'
    post:
      tags:
        - Loja
      summary: Criando Uma Loja
      description: Criando Uma Loja na Plataforma
      operationId: criandoLoja
      requestBody:
        description: Criando nova Loja
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LojaApiRequest'
        required: true
      responses:
        '201':
          description: created successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LojaApiResponse'
        '400':
          description: bad request operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 400 BAD_REQUEST
                  erros:
                    type: array
                    items:
                      properties:
                        mensagem:
                          type: string
                          example: "não deve ser nulo"
                        campo:
                          type: string
                          example: "cnpj"
        '422':
          description: unprocessable operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 422 UNPROCESSABLE_ENTITY
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: CNPJ já Cadastrado.
                        mensagem:
                          type: string
                          example: "CNPJ Já cadastrado."
  /loja/{guid}:
    get:
      tags:
        - Loja
      summary: Buscando Loja Pelo GUID
      description: Buscando Loja Pelo GUID na Plataforma
      operationId: buscandoLojaPeloGUID
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LojaApiResponse'
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: "Loja não encontrada."
    put:
      tags:
        - Loja
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          example: aaaa-bbbb-cccc-dddd-eeeee
          required: true
          schema:
            type: string
      summary: Alterando uma Loja
      description: Alterando uma Loja na Plataforma
      operationId: alterandoLoja
      requestBody:
        description: Alterando Loja Existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LojaUpdateApiRequest'
      responses:
        '202':
          description: accepted successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LojaApiResponse'
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: "Loja não encontrada."
    delete:
      tags:
        - Loja
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          required: true
          schema:
            type: string
      summary: Apagando uma Loja
      description: Apagando uma Loja na Plataforma
      operationId: apagandoLoja
      responses:
        '204':
          description: No Content successful operation
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: "Loja não encontrada."
        '422':
          description: unprocessable operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 422 UNPROCESSABLE_ENTITY
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não permitido Excluir.
                        mensagem:
                          type: string
                          example: "Loja não pode ser Excluida. Pois está associada a alguma promoção."
    patch:
      tags:
        - Loja
      parameters:
        - name: guid
          in: path
          description: Informe o GUID
          required: true
          schema:
            type: string
      summary: Alterando Status de uma Loja
      description: Alterando Status de uma Loja na Plataforma
      operationId: alterandoStatusLoja
      requestBody:
        description: Alterando Loja Existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusApiRequest'
        required: true
      responses:
        '202':
          description: Accepted successful operation
        '404':
          description: not found operation
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: 404 NOT_FOUND
                  erros:
                    type: array
                    items:
                      properties:
                        codigo:
                          type: string
                          example: Não encontrado.
                        mensagem:
                          type: string
                          example: "Loja não encontrada."
  /loja/paginada:
    get:
      tags:
        - Loja
      summary: Buscando Lista de Loja Pagina
      description: Buscando Lista de Loja pagina na Plataforma
      operationId: buscandoListaPaginadaLoja
      parameters:
        - name: page
          description: "Numero da pagina atual"
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          description: "Quantidade de registro por pagina"
          in: query
          schema:
            type: integer
            default: 10
        - name: cnpj
          description: "CNPJ da Loja"
          in: query
          schema:
            type: string
            default: ""
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginacaoApiResponse'
                  - type: object
                    properties:
                      lojas:
                        type: array
                        items:
                          $ref: '#/components/schemas/LojaApiResponse'
  /loja/ativa:
    get:
      tags:
        - Loja
      summary: Buscando Lista de Loja Ativa
      description: Buscando Lista de Loja Ativa na Plataforma
      operationId: buscandoListaLojaAtiva
      parameters:
        - name: email
          description: "E-mail do usuario logado"
          example: "mario7lagoas@gmail.com"
          in: query
          schema:
            type: string
            default: ""
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LojaApiResponse'
  /loja/print:
    post:
      tags:
        - Loja
      summary: Relatorio de Loja
      description: Relatorio de Loja na Plataforma
      operationId: relatorioLoja
      requestBody:
        description: Imprimindo um relatorio de  Loja
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/LojaApiResponse'
      responses:
        '200':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                example: "data:application/pdf;base64,JVBERi0xLjUKJeLjz9MKMyAwIG9iago8PC9Db2xvclNwYWNlL0RldmljZVJHQi9TdWJ0eXBlL0ltYWdlL0hlaWdodCAzNTQvRmlsdGVyL0RDVERlY29kZS9UeXBlL1hPYmplY3QvV2lkdGggMjQ4MC9CaXRzUGVyQ29tcG9uZW50IDgvTGVuZ3RoIDM2ODg5Pj5zdHJlYW0K/9j/4AAQSkZJRgABAgAAZABkAAD/7AARRHVja3kAAQAEAAAACgAA/+4ADkFkb2JlAGTAAAAAAf/bAIQAFBAQGRIZJxcXJzImHyYyLiYmJiYuPjU1NTU1PkRBQUFBQUFERERERERERERERERERERERERERERERERERERERAEVGRkgHCAmGBgmNiYgJjZENisrNkREREI1QkRERERERERERERERERERERERERERERERERERERERERERERERERE/8AAEQgBYgmwAwEiAAIRAQMRAf/..."
...
components:
  schemas:
    PaginacaoApiResponse:
      properties:
        number:
          type: integer
          example: 0
          description: "Pagina atual"
        totalPages:
          type: integer
          example: 1
          description: "Total de paginas"
        totalElements:
          type: integer
          example: 1
          description: "Total de Items"
...
    UpdateStatusApiRequest:
      required:
        - status
      type: object
      properties:
        status:
          type: boolean
          description: "Status [true/false]"
          example: true
...
    LojaApiRequest:
      required:
        - nome
        - identificacao
        - cnpj
        - status
      type: object
      properties:
        nome:
          type: string
          description: "Nome da Loja"
          example: "Loja 001"
        identificacao:
          type: string
          description: "Identificação da Loja"
          example: "0001"
        cnpj:
          type: string
          description: "CNPJ da Loja"
          example: "42492701000135"
        status:
          type: boolean
          description: "Status da Loja"
          example: true
    LojaUpdateApiRequest:
      type: object
      properties:
        nome:
          type: string
          description: "Nome da Loja"
          example: "Loja 001"
        identificacao:
          type: string
          description: "Identificação da Loja"
          example: "0001"
        cnpj:
          type: string
          description: "CNPJ da Loja"
          example: "42492701000135"
        status:
          type: boolean
          description: "Status da Loja"
          example: true
    LojaApiResponse:
      properties:
        guid:
          type: string
          example: "aaaa-bbbb-cccc-dddd-eeeee"
          description: "Identificador guid"
        nome:
          type: string
          example: "Loja 001"
          description: "Nome da Loja"
        cnpj:
          type: string
          example: "4249270100013"
          description: "CNPJ da Loja"
        identificacao:
          type: string
          example: "0001"
          description: "Idenficação da Loja"
        status:
          type: boolean
          example: true
          description: "Status da Loja"
        dataCadastro:
          type: string
          example: "2024-05-07 11:48:11"
          description: "Data do cadastro"
...